Make 64 bit clean This was quite a nightmare. Main reason: clash between use of "int" and array lengths (which are size_t). Implicit conversion of wider integer types to narrower ones has been disallowed recently in D2. Whenever you assign a size_t to a int, you need a cast. This was especially painful with integer indices. ("int" was mostly only used when small array lengths were implied, especially in conjunction with strings. So there were no real bugs, probably.)
There were some issues with C functions taking "int" for sizes, too.

+++ src/common/globalconsole.d
@@ -163,7 +163,7 @@ private void onActivate(MyBox[] args, Output output) {
     activateConsole();
 }
 
-private void onTabComplete(cstring line, int cursor1, int cursor2,
+private void onTabComplete(cstring line, size_t cursor1, size_t cursor2,
     EditDelegate edit)
 {
     //xxx: somehow should call CommandLine tabhandler if the line starts with
